pyproject.toml
[project]
name = "safco-agent"
version = "0.1.0"
description = "Agent-based scraper for Safco Dental Supply product catalog"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Yuchen Lin" }]
dependencies = [
    "httpx[http2]>=0.27",
    "playwright>=1.45",
    "selectolax>=0.3.21",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "structlog>=24.1",
    "tenacity>=8.4",
    "aiolimiter>=1.1",
    "typer>=0.12",
    "rich>=13.7",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "anthropic>=0.34",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
]

[project.scripts]
safco = "safco_agent.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/safco_agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
